Reorganizes project structure and updates documentation
