+++ archive/PhotoOrganizer_v2/PhotoOrganizerScript_original.ps1
