+++ assets/icons/camera/about.txt
