+++ assets/icons/camera/android-chrome-512x512.png
