+++ assets/icons/combine icons.py
