--- bad prototypes/power automate poging/PowerAutomate_PhotoOrganizer.ps1
@@ -1,1 +1,0 @@
-﻿PathToScrip -source "PathToSource" -dest "PathToDestination" -format "DateFormat"
